ci: Avoid fetching another remote if operating in the GNOME namespace
Copy of a patch by Emmanuele Bassi in GTK
(https://gitlab.gnome.org/GNOME/gtk/-/commit/6bdca276a2318fa1fd2f9f9df7a475c52bb44a33)
which avoids pulling the upstream remote if the branch being tested is
already in the GNOME namespace (in which case, `origin` is already the
remote we want).

This will reduce git pulls during CI if running on an MR in that
namespace.

Signed-off-by: Philip Withnall <withnall@endlessm.com>